Split gem module specs along module boundaries
