+++ spec/gem/state_spec.rb
@@ -1,0 +1,10 @@
+#
+# Licensed under the MIT license
+# Copyright (C) 2013-2014 Red Hat, Inc.
+
+require 'polisher/gem'
+
+module Polisher
+  describe Gem do
+  end # describe Gem
+end # module Polisher
